created typing component and page, moved things
