+++ src/lib/mysql.tsx
@@ -1,0 +1,11 @@
+import mysql from 'mysql2/promise'
+
+const pool = mysql.createPool({
+    host: process.env.DB_HOST,
+    user: process.env.DB_USER,
+    password: process.env.DB_PASS,
+    database: process.env.DB_DATABASE,
+    waitForConnections: true
+})
+
+export default pool
